Update README and add tests for deprecate_public usage on Module in Ruby 3
With the changes in Ruby 3, you can now use deprecate_public on
a module after the module has been included in a class.  Note that
deprecate_public_constant oon a module after inclusion in a class
doesn't work (that calls extend on the module, not the classes
including the module).